refactor: add dedicated limits and key files
