+++ internal/rts/args.go
@@ -15,8 +15,65 @@ func NewArgs(ctx *Ctx, pos Pos, args []Value, sig string) Args {
 	return Args{ctx: ctx, pos: pos, args: args, sig: sig}
 }
 
+func argCount(ctx *Ctx, pos Pos, args []Value, want int, sig string) error {
+	if len(args) != want {
+		return Errf(ctx, pos, "%s expects %d args", sig, want)
+	}
+	return nil
+}
+
+func argCountRange(ctx *Ctx, pos Pos, args []Value, min, max int, sig string) error {
+	if len(args) < min || len(args) > max {
+		return Errf(ctx, pos, "%s expects %d-%d args", sig, min, max)
+	}
+	return nil
+}
+
+func strArg(ctx *Ctx, pos Pos, v Value, sig string) (string, error) {
+	s, err := ToStr(ctx, pos, v)
+	if err != nil {
+		return "", err
+	}
+	if err := CheckStr(ctx, pos, s); err != nil {
+		return "", err
+	}
+	return s, nil
+}
+
+func numArg(ctx *Ctx, pos Pos, v Value, sig string) (float64, error) {
+	if v.K != VNum {
+		return 0, Errf(ctx, pos, "%s expects number", sig)
+	}
+	return v.N, nil
+}
+
+func listArg(ctx *Ctx, pos Pos, v Value, sig string) ([]Value, error) {
+	if v.K == VNull {
+		return nil, nil
+	}
+
+	if v.K != VList {
+		return nil, Errf(ctx, pos, "%s expects list", sig)
+	}
+
+	if err := CheckList(ctx, pos, len(v.L)); err != nil {
+		return nil, err
+	}
+	return v.L, nil
+}
+
+func dictArg(ctx *Ctx, pos Pos, v Value, sig string) (map[string]Value, error) {
+	if v.K == VNull {
+		return nil, nil
+	}
+	if v.K != VDict {
+		return nil, Errf(ctx, pos, "%s expects dict", sig)
+	}
+	return v.M, nil
+}
+
 func (a Args) Count(want int) error {
-	return ArgCount(a.ctx, a.pos, a.args, want, a.sig)
+	return argCount(a.ctx, a.pos, a.args, want, a.sig)
 }
 
 func (a Args) None() error {
@@ -24,7 +81,7 @@ func (a Args) None() error {
 }
 
 func (a Args) CountRange(min, max int) error {
-	return ArgCountRange(a.ctx, a.pos, a.args, min, max, a.sig)
+	return argCountRange(a.ctx, a.pos, a.args, min, max, a.sig)
 }
 
 func (a Args) Len() int {
@@ -40,7 +97,7 @@ func (a Args) Arg(i int) Value {
 }
 
 func (a Args) Str(i int) (string, error) {
-	return StrArg(a.ctx, a.pos, a.args[i], a.sig)
+	return strArg(a.ctx, a.pos, a.args[i], a.sig)
 }
 
 func (a Args) ToStr(i int) (string, error) {
@@ -48,7 +105,7 @@ func (a Args) ToStr(i int) (string, error) {
 }
 
 func (a Args) Num(i int) (float64, error) {
-	return NumArg(a.ctx, a.pos, a.args[i], a.sig)
+	return numArg(a.ctx, a.pos, a.args[i], a.sig)
 }
 
 func (a Args) FiniteNum(i int) (float64, error) {
@@ -75,11 +132,11 @@ func (a Args) ScalarStr(i int) (string, error) {
 }
 
 func (a Args) List(i int) ([]Value, error) {
-	return ListArg(a.ctx, a.pos, a.args[i], a.sig)
+	return listArg(a.ctx, a.pos, a.args[i], a.sig)
 }
 
 func (a Args) Dict(i int) (map[string]Value, error) {
-	return DictArg(a.ctx, a.pos, a.args[i], a.sig)
+	return dictArg(a.ctx, a.pos, a.args[i], a.sig)
 }
 
 func (a Args) Key(i int) (string, error) {
