+++ internal/rts/map.go
@@ -1,6 +1,7 @@
 package rts
 
 import (
+	"fmt"
 	"maps"
 	"strings"
 )
@@ -23,7 +24,7 @@ type ms struct {
 func newMS(n string) ms {
 	return ms{
 		g: n + ".get(name)",
-		h: n + ".Has(name)",
+		h: n + ".has(name)",
 		r: n + ".require(name[, msg])",
 	}
 }
@@ -34,13 +35,13 @@ func lowerMap(src map[string]string) map[string]string {
 	}
 	out := make(map[string]string, len(src))
 	for k, v := range src {
-		out[lowerKey(k)] = v
+		out[lookupKey(k)] = v
 	}
 	return out
 }
 
 func mapLookup(m map[string]string, name string) (string, bool) {
-	v, ok := m[lowerKey(name)]
+	v, ok := m[lookupKey(name)]
 	return v, ok
 }
 
@@ -65,10 +66,11 @@ func mapIndex(m map[string]string, key Value) (Value, error) {
 }
 
 func mapGet(ctx *Ctx, pos Pos, args []Value, sig string, m map[string]string) (Value, error) {
-	if err := ArgCount(ctx, pos, args, 1, sig); err != nil {
+	na := NewArgs(ctx, pos, args, sig)
+	if err := na.Count(1); err != nil {
 		return Null(), err
 	}
-	k, err := Key(pos, args[0])
+	k, err := Key(pos, na.Arg(0))
 	if err != nil {
 		return Null(), WrapErr(ctx, err)
 	}
@@ -80,10 +82,11 @@ func mapGet(ctx *Ctx, pos Pos, args []Value, sig string, m map[string]string) (V
 }
 
 func mapHas(ctx *Ctx, pos Pos, args []Value, sig string, m map[string]string) (Value, error) {
-	if err := ArgCount(ctx, pos, args, 1, sig); err != nil {
+	na := NewArgs(ctx, pos, args, sig)
+	if err := na.Count(1); err != nil {
 		return Null(), err
 	}
-	k, err := Key(pos, args[0])
+	k, err := Key(pos, na.Arg(0))
 	if err != nil {
 		return Null(), WrapErr(ctx, err)
 	}
@@ -98,10 +101,11 @@ func mapRequire(
 	sig, obj string,
 	m map[string]string,
 ) (Value, error) {
-	if err := ArgCountRange(ctx, pos, args, 1, 2, sig); err != nil {
+	na := NewArgs(ctx, pos, args, sig)
+	if err := na.CountRange(1, 2); err != nil {
 		return Null(), err
 	}
-	k, err := KeyArg(ctx, pos, args[0], sig)
+	k, err := na.Key(0)
 	if err != nil {
 		return Null(), err
 	}
@@ -110,4 +114,27 @@ func mapRequire(
 		return Str(v), nil
 	}
 	return Null(), reqErr(ctx, pos, obj, k, args)
+}
+
+func reqMsg(ctx *Ctx, pos Pos, args []Value) (string, error) {
+	if len(args) < 2 {
+		return "", nil
+	}
+
+	s, err := ToStr(ctx, pos, args[1])
+	if err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(s), nil
+}
+
+func reqErr(ctx *Ctx, pos Pos, obj, key string, args []Value) error {
+	msg, err := reqMsg(ctx, pos, args)
+	if err != nil {
+		return err
+	}
+	if msg == "" {
+		msg = fmt.Sprintf("missing required %s: %s", obj, key)
+	}
+	return Errf(ctx, pos, "%s", msg)
 }
