+++ internal/rts/vars.go
@@ -90,22 +90,23 @@ func (o *varsObj) requireFn(ctx *Ctx, pos Pos, args []Value) (Value, error) {
 
 func (o *varsObj) setFn(ctx *Ctx, pos Pos, args []Value) (Value, error) {
 	sig := o.name + ".set(name, value)"
-	if err := ArgCount(ctx, pos, args, 2, sig); err != nil {
+	na := NewArgs(ctx, pos, args, sig)
+	if err := na.Count(2); err != nil {
 		return Null(), err
 	}
 	if o.mut == nil {
 		return Null(), Errf(ctx, pos, "%s is read-only", o.name)
 	}
-	name, err := KeyArg(ctx, pos, args[0], sig)
+	name, err := na.Key(0)
 	if err != nil {
 		return Null(), err
 	}
-	val, err := ScalarStr(ctx, pos, args[1], sig)
+	val, err := na.ScalarStr(1)
 	if err != nil {
 		return Null(), err
 	}
 	o.mut.SetVar(name, val)
-	key := lowerKey(name)
+	key := lookupKey(name)
 	o.m[key] = val
 	return Null(), nil
 }
@@ -170,25 +171,26 @@ func (o *globalObj) setFn(ctx *Ctx, pos Pos, args []Value) (Value, error) {
 		}
 	}
 	o.mut.SetGlobal(name, val, secret)
-	key := lowerKey(name)
+	key := lookupKey(name)
 	o.m[key] = val
 	return Null(), nil
 }
 
 func (o *globalObj) delFn(ctx *Ctx, pos Pos, args []Value) (Value, error) {
 	sig := o.name + ".delete(name)"
-	if err := ArgCount(ctx, pos, args, 1, sig); err != nil {
+	na := NewArgs(ctx, pos, args, sig)
+	if err := na.Count(1); err != nil {
 		return Null(), err
 	}
 	if o.mut == nil {
 		return Null(), Errf(ctx, pos, "%s is read-only", o.name)
 	}
-	name, err := KeyArg(ctx, pos, args[0], sig)
+	name, err := na.Key(0)
 	if err != nil {
 		return Null(), err
 	}
 	o.mut.DelGlobal(name)
-	key := lowerKey(name)
+	key := lookupKey(name)
 	delete(o.m, key)
 	return Null(), nil
 }
